requirements.txt
django-rest-knox
djangorestframework
django
